Updated scoreData from Object to Array of Objects
Previous method of storing player data within an object, with the key
being their name, was not going to be scalable. For one, it meant that
no two players could share an id. It also meant that changing the name
of the player would be very convoluted.
Instead, now the data is an array of objects, and the position within
that array is passed to each scorecard, so that it can pass it to
whatever update function it's calling, and that index can be used to
update the correct data object.