Added Variable Increment Values
Closes #2
There is now an increment property in each player object, and that is
the value used to increment/decrement player scores. For now, they are
not tied to each other in any way, and the function for changing the
values reflects this, but with how I imagine using the feature in the
future it will just be a global setting rather than each player (why
would certain players have different value?).
Other option is to have it both ways, and let the user choose if they
want them tied together, or able to be separate.