Changing Player Name Actually Changes It
Closes #6
By passing the function that update the player data to the player
screen, it is able to take advantage of it and have its changes reflect
locally on that player screen, as well as updating the information on
the score screen. For now it is just the name I got working.