Added Global Provider for Global Settings
Score data should be used for strictly the players and their data, so I
made another context file that will be used to hold all of the data
about the app/game in general. Also added the provider to the
UserProvider, so that I didn't have to update that call.