Added Ability to Toggle High/Low Score Wins
Some games, you want the lower score to be the winner, so added global
settings that would keep track of if the high score wins or not. Added
this context to the NavBar, and updated the winner check so that it will
compare to a really negative/positive depending on that value.