Changes to Player Data Only Save on Button Click
You can now update the values for a player, without it constantly
updating the state. Now, in order to save all of the states globally,
the save button must be clicked. If it isn't the changes aren't saved,
and the player is brought back to the score screen with their original
data intact.