Most of Global Settings Working
Able to update things within player screen again, and began work on
getting global settings to work. The current thing not working is
increment, since I'm using the scoreDataDispatch and it doesn't have
access to globalData. Will either have to add it in for increment, or
rework the entire global settings pipeline.