Cleaned Up Score Card Appearance
Centered items, made the edit button go all the way to the right and
added a spacer view in between it and the name, expanded the size of the
increment pressables, and added the option to change the hue of the
card. The hue given will be the hue of the card, and the title will be
the same hue, but a higher lightness.