Added Some Game Specific Settings
Closes #12
Added a few different games and their settings: rummy, euchre, cribbage,
scopa, and frolf. Almost all are standard increment, resetValue, and
highestScoreWins. However, I also added a winningScore and
pointReasonArray property. winningScore is pretty self explanatory, the
pointReasonArray I'm thinking could be useful for keeping a more
accurate log. Specifically, I used it in Scopa, where the reasons for
gaining points can be 7 of diamonds, most diamonds/cards, or best set of
suits. I figured it would be a pop up or something similar when the
increment button is clicked, to quickly select the reason, and then the
log could say "Player 1 gained 1 point for Most Diamonds", or something
similar.