Displaying Current Game, Storing it in Global Data
Closes #16
On the bottom nav the current game mode is displayed, and this current
game mode is stored within the currentGame property of globalData. The
NewGameScreen component has been updated to save the selected game if
the save button (also new) is clicked, and doesn't save if you just
navigate back.