Updated globalDataDispatch Update Behaviour
Closes #17
Before it was taking a value/key pair, or value/key array pairs, and you
had to specifically call out if you were updating one or multiple. Now,
you can just pass it an object with the new settings, and it will spread
it onto the current data to update what needs to be, and now it is just
"update", no need to specify multiple or not.