Allowed Editing of Game Settings (Local Only)
Closes #18
Player can edit the settings of a game when creating a new game.
Everything appears to be working fine with not converting numbers to
strings, and keeping higherScoreWins as a boolean.

For the future, I will want to have the boolean values be a different
Component (just a switch), and I need to update it so that changing a
game of the same type already being played will still update global
settings, but not reset all players.