Allowed Updating of Current Game Settings
Previous behaviour prevented game settings changes being emitted if the
game being edited was the current game. This was done to stop the game
from being reset if editing the same game you have now. Unfortunately,
that also meant you would have to create a new game, and then go back
and create the game you wanted to in the first place. So now, editing
the same game will just update the settings, without resetting the
players to resetValue. This will obviously be an issue when trying to
actually create a new game of the current one being played, will have to
come up with a pop up to see if they want to reset or not.