Created Default Color Theme
Closes #13
Came up with a format that I think will work well for the color themes,
and added a default one. It appears to be fully working with the score
cards. Also, added a bit to the ScoreCard component in order to have the
title be slightly lighter, as well as changing text from white to black
depending on how light the color is.